feat: add ProtocolConfig types and fix clippy/test issues (#115) (#135)
- Add ProtocolConfig interface and parseProtocolConfig() to @agenc/runtime
- Add comprehensive type guard validating all required fields
- Add range validation for disputeThreshold (1-100), protocolFeeBps (<=10000), slashPercentage (0-100)
- Add protocol.test.ts with 31 tests covering parsing, validation, and edge cases
- Fix clippy warnings: remove crate-level too_many_arguments allow, add function-level allows only where needed
- Fix test reliability: ensure agents registered in beforeEach hook

+++ programs/agenc-coordination/src/instructions/resolve_dispute.rs
@@ -107,10 +107,11 @@ pub fn handler(ctx: Context<ResolveDispute>) -> Result<()> {
     require!(total_votes > 0, CoordinationError::InsufficientVotes);
 
     // Calculate approval percentage
-    let approval_pct = (dispute.votes_for as u64)
+    let approval_pct = dispute
+        .votes_for
         .checked_mul(PERCENT_BASE)
         .ok_or(CoordinationError::ArithmeticOverflow)?
-        .checked_div(total_votes as u64)
+        .checked_div(total_votes)
         .ok_or(CoordinationError::ArithmeticOverflow)? as u8;
 
     // Determine outcome based on threshold
